% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{approx_cpp}
\alias{approx_cpp}
\title{Linear interpolation function}
\usage{
approx_cpp(x, y, xout)
}
\arguments{
\item{x}{NumericVector of x coordinates of points to be interpolated}

\item{y}{NumericVector of y coordinates of points to be interpolated}

\item{xout}{NumericVector of points at which to interpolate}
}
\value{
NumericVector of interpolated values
}
\description{
This function carries out linear interpolation, reproducing a subset of the behaviour of stats::approx in R.
}
\details{
Linear interpolation is performed with f(x)=f(x0)+[(f(x1)-f(x0))/(x1-x0)]*(x-x0). x is assumed to be sorted; if sorted in descending order, the order of elements in x and y is reversed. If a value of xout lies outside the range [min(x), max(x)], the value is interpolated using the nearest extremum (this corresponds to the behaviour of stats::approx with method=2).
}
\author{
Tom Willis
}
