% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bh}
\alias{bh}
\title{bh}
\usage{
bh(P, alpha)
}
\arguments{
\item{P}{list of p-values}

\item{alpha}{FDR control level}
}
\value{
indices of p-values to reject
}
\description{
Run the Benjamini-Hochberg procedure
}
\examples{
# no
}
