% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cecdf}
\alias{cecdf}
\title{conditional eCDF
calculate p(X<x|Y<y) for fixed vectors of points x and y}
\usage{
cecdf(x, y)
}
\description{
conditional eCDF
calculate p(X<x|Y<y) for fixed vectors of points x and y
}
