% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cFDR.R
\name{cFDR.assess}
\alias{cFDR.assess}
\title{cFDR}
\usage{
cFDR.assess(Z, Zc, alt, alpha, Z2 = NULL, Zc2 = NULL, ...)
}
\arguments{
\item{Z}{vector of Z scores}

\item{Zc}{vector of Z scores for the conditional tests}

\item{alt}{logical vector of equal length with Z, with TRUE indicating alternative hypotheses}

\item{alpha}{FDR threshold for assessment}

\item{...}{other arguments passed to cFDR}
}
\value{
named vector of requested alpha, estimated alpha, power, fdr
}
\description{
Assess power and FDR control for the *independent* cFDR procedure for simulated
data where it is known whether each Z score corresponds to the
null or alternative hypothesis
}
\author{
Chris Wallace
}
