% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cfdr}
\alias{cfdr}
\title{cfdr}
\usage{
cfdr(p, q, sub = which(qnorm(p/2)^2 + qnorm(q/2)^2 > 4), exclude = NULL,
  adj = F)
}
\arguments{
\item{p}{vector of p-values for dependent variable of interest}

\item{q}{vector of p-values from other dependent variable#}

\item{sub}{list of indices at which to compute cFDR estimates}

\item{exclude}{list of indices to exclude (each point (p[i],q[i]) is still automatically included in the computation of its own cFDR value)}

\item{adj}{include estimate of Pr(H^p=0|Q<q) in estimate}
}
\value{
vector of cFDR values; set to 1 if index is not in 'sub'
}
\description{
Estimate cFDR at a set of points using counting-points method (cFDR1 or cFDR1s)
}
\examples{
# Generate standardised simulated dataset
set.seed(1); n=10000; n1p=100; n1pq=100; n1q=100
zp=c(rnorm(n1p,sd=3), rnorm(n1q,sd=1),rnorm(n1pq,sd=3), rnorm(n-n1p-n1q-n1pq,sd=1))
zq=c(rnorm(n1p,sd=1), rnorm(n1q,sd=3),rnorm(n1pq,sd=3), rnorm(n-n1p-n1q-n1pq,sd=1))
p=2*pnorm(-abs(zp)); q=2*pnorm(-abs(zq))

cx=cfdr(p,q)

plot(p,q,cex=0.5,xlim=c(0,0.05)); points(p[which(cx<0.5)],q[which(cx<0.5)],col="red") 
}
\author{
James Liley
}
