% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cFDR.R
\name{cFDR2}
\alias{cFDR2}
\title{cFDR}
\usage{
cFDR2(Z1, Z2, Zc1, Zc2, eps = 0.01, alpha = seq(0.01, 0.1, by = 0.005),
  do.optimise = FALSE, ...)
}
\arguments{
\item{eps}{tolerance for FDR control to alpha}

\item{alpha}{vector of thresholds. overall cFDR will be assessed
if all tests with cFDR<alpha were rejected}

\item{do.optimise}{logical, default FALSE. if true, and a single
value of alpha is supplied, the rejection threshold will be
optimised to control the overall FDR at alpha.}

\item{...}{other arguments passed to cFDR.tune}
}
\value{
a named list with three components.  alpha is the
    threshold for rejection, est.cFDR is the estimated FDR at this
    threshold, cFDR is a vector of conditional FDRs in rectangles
    defined the corresponding absolute Z and Zc values.
}
\description{
Calculate cFDR given vectors of *independent* test and conditional Z scores
}
\author{
Chris Wallace
}
