% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cFDR.R
\name{depcFDR}
\alias{depcFDR}
\title{depcFDR}
\usage{
depcFDR(Z, Zc, rho, mu = 0, muc, eps = 0.01, p.max = 0.1,
  alpha = seq(0.01, 0.1, by = 0.005), do.optimise = FALSE, ...)
}
\arguments{
\item{Z}{vector of Z scores}

\item{rho}{correlation between Z and Zc under the null hypothesis}

\item{alpha}{FDR threshold for assessment}

\item{...}{other arguments passed to cFDR.tune}
}
\value{
a named list with three components.  alpha is the
    threshold for rejection, est.cFDR is the estimated FDR at this
    threshold, cFDR is a vector of conditional FDRs in rectangles
    defined the corresponding absolute Z and Zc values.
}
\description{
Calculate cFDR given vectors of *dependent* test and conditional Z scores
}
\author{
Chris Wallace
}
