% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cFDR.R
\name{FDR.assess}
\alias{FDR.assess}
\title{assess power and FDR control of FDR procedure}
\usage{
FDR.assess(Z, alt, alpha)
}
\arguments{
\item{Z}{vector of Z scores}

\item{alt}{logical vector of equal length with Z, with TRUE indicating alternative hypotheses}

\item{alpha}{FDR threshold for assessment}
}
\value{
named vector of alpha, power, FDR
}
\description{
Assess power and FDR control for simulated data where it is known whether each Z score corresponds to the null or alternative hypothesis
}
\details{
.. content for \details{} ..
}
\author{
Chris Wallace
}
