% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fit.2g.parallel}
\alias{fit.2g.parallel}
\title{fit.2g.parallel}
\usage{
fit.2g.parallel(
  P,
  pars = c(0.5, 1.5),
  weights = rep(1, min(length(Z), dim(Z)[1])),
  sigma_range = c(1, 100),
  rho = 0,
  ncores = 1,
  ...
)
}
\arguments{
\item{P}{numeric vector of observed data, either p-values or z-scores. If rho=0, should be one-dimensional vector; if rho is set, should be bivariate observations (P,Q)}

\item{pars}{initial values for parameters}

\item{weights}{optional weights for parameters}

\item{sigma_range}{range of possible values for sigma (closed interval). Default [1,100]}

\item{rho}{correlation between P and Q if bivariate observations are specified}

\item{ncores}{number of cores on which to run parallel optimisation procedure}

\item{...}{arguments passed to optimParallel call}
}
\value{
a list containing parameters pars, likelihoods under h1 (Z distributed as above), likelihood under h0 (Z~N(0,1)) and likelihood ratio lr.
}
\description{
Fit a specific two Gaussian mixture distribution to a set of P or Z values using a parallel implementation of optim.
}
\details{
Assumes 
Z ~ N(0,1) with probability pi0, Z ~ N(0,sigma^2) with probability 1-pi0

Returns MLE for pi0 and sigma. Uses R's optim function. Can weight observations.
}
\author{
James Liley
}
