% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fit.2g}
\alias{fit.2g}
\title{fit.2g}
\usage{
fit.2g(P, pars = c(0.5, 1.5), weights = rep(1, min(length(Z), dim(Z)[1])),
  sigma_range = c(1, 100), rho = 0, ...)
}
\arguments{
\item{P}{numeric vector of observed data, either p-values or z-scores. If rho=0, should be one-dimensional vector; if rho is set, should be bivariate observations (P,Q)}

\item{pars}{initial values for parameters}

\item{weights}{optional weights for parameters}

\item{sigma_range}{range of possible values for sigma (closed interval). Default [1,100]}
}
\value{
a list containing parameters pars, likelihoods under h1 (Z distributed as above), likelihood under h0 (Z~N(0,1)) and likelihood ratio lr.
}
\description{
Fit a specific two Guassian mixture distribution to a set of P or Z values.
}
\details{
Assumes 
Z ~ N(0,1) with probability pi0, Z ~ N(0,sigma^2) with probability 1-pi0

Returns MLE for pi0 and sigma. Uses R's optim function. Can weight observations.
}
\examples{
sigma=2; pi0 <- 0.8

n=10000; n0=round(pi0*n); n1=n-n0
Z = c(rnorm(n0,0,1),rnorm(n1,0,sqrt(1+ (sigma^2))))
fit=fit.2g(Z)
fit$pars
}
\author{
James Liley
}
