% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-normals.R
\name{fit.vars}
\alias{fit.vars}
\title{fit.em}
\usage{
fit.vars(Z, pi0 = 0.9, s2 = 1, tol = 1e-04, verbose = TRUE,
  maxit = 10000)
}
\arguments{
\item{Z}{numeric vector of observed data}

\item{pi0}{initial proportion of samples from f0}

\item{s2}{initial value for sigma^2}

\item{tol}{how small a change lhood prompts continued optimization}

\item{maxit}{maximum number of iterations}
}
\value{
a list of two objects giving fitted values and history
}
\description{
Fit a specific two Guassian mixture distribution
}
\details{
Assumes Z ~ N(0,1) with prob pi0,
Z ~ N(0,1+sigma^2) with prob 1-pi0

Aims to estimate sigma^2 and pi0.
}
\examples{
s2 <- 10
pi0 <- 0.8
n <- 10000
Z <- c(rnorm(round(n*pi0),0,1),rnorm(round(n*(1-pi0)),0,sqrt(1+s2)))
fit<-fit.em(Z)
fit$pars
fit$history
}
\references{
Cut down from more general version at
https://gist.github.com/chr1swallace/11421212 for speed.
}
\author{
Chris Wallace
}
