% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vl_mode2}
\alias{vl_mode2}
\title{Returns coordinates of L-regions for cFDR method. Drop-in replacement for the cfdr::vl function in the use case wherein indices and fold are supplied and mode=2.}
\usage{
vl_mode2(
  p,
  q,
  indices,
  fold,
  adj = TRUE,
  at = NULL,
  nt = 5000L,
  nv = 1000L,
  p_threshold = 0,
  scale = as.character(c("p", "z")),
  closed = TRUE,
  verbose = FALSE,
  gx = 1e-05
)
}
\arguments{
\item{p}{principal p-values}

\item{q}{conditional p-values}

\item{indices}{indices of points at which to compute v(L)}

\item{fold}{indices of points to exclude when from calculation of L-curves}

\item{adj}{adjust cFDR values and hence curves L using estimate of Pr(H0|Pj<pj)}

\item{at}{cfdr cutoff/cutoffs. Defaults to null}

\item{nt}{number of test points in x-direction, default 5000}

\item{nv}{resolution for constructing L-curves, default 1000}

\item{p_threshold}{if H0 is to be rejected automatically whenever p<p_threshold, include this in all regions L}

\item{scale}{return curves on the p- or z- plane. Y values are equally spaced on the z-plane.}

\item{closed}{determines whether curves are closed polygons encircling regions L (closed=T), or lines indicating the rightmost border of regions L}

\item{verbose}{print progress}
}
\value{
list containing elements x, y. Assuming n curves are calculated (where n=length(indices) or length(at)) and closed=T, x is a matrix of dimension n x (4+nv), y ix a vector of length (4+nv).
}
\description{
Returns coordinates of L-regions for cFDR method. Drop-in replacement for the cfdr::vl function in the use case wherein indices and fold are supplied and mode=2.
}
\author{
Tom Willis
}
