% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{wecdf}
\alias{wecdf}
\title{weighted eCDF
calculate p(X<x)(w) yfor fixed vectors of points x and weights w}
\usage{
wecdf(x, w)
}
\description{
weighted eCDF
calculate p(X<x)(w) yfor fixed vectors of points x and weights w
}
